Addressed changes
