+++ state_manage/reducers/index.ts
@@ -1,0 +1,6 @@
+import { combineReducers } from 'redux';
+import profileScreenReducer from './profileScreenReducer'
+
+export default combineReducers({
+    profile: profileScreenReducer
+})
