+++ state_manage/store.ts
